[refactor] clean code
