+++ TelemetryDevice/Services/Sniffer/IPacketSniffer.cs
@@ -1,11 +1,10 @@
 ﻿namespace TelemetryDevices.Services.Sniffer
 {
-    public interface IPacketSniffer
+    public interface IPacketSniffer : IDisposable
     {
         public event Action<byte[], int> PacketReceived;
         public void AddPort(int port);
         public void RemovePort(int port);
         public List<int> GetPorts();
-        public void Dispose();
     }
 }
