+++ TelemetryDevice/TelemetryDevices.csproj
@@ -19,10 +19,7 @@
     <PackageReference Include="Newtonsoft.Json.Bson" Version="1.0.3" />
     <PackageReference Include="PacketDotNet" Version="1.4.8" />
     <PackageReference Include="SharpPcap" Version="6.3.1" />
-    <PackageReference
-      Include="System.Threading.Tasks.Dataflow"
-      Version="10.0.0-preview.6.25358.103"
-    />
+    <PackageReference Include="System.Threading.Tasks.Dataflow" Version="10.0.0-preview.6.25358.103" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\..\Core\Core\Core.csproj" />
